+++ api/package.json
@@ -30,5 +30,11 @@
 		"typescript": "^5.0.0",
 		"typescript-eslint": "^8.20.0",
 		"vite": "^7.0.4"
+	},
+	"dependencies": {
+		"@supabase/ssr": "^0.6.1",
+		"@supabase/supabase-js": "^2.52.1",
+		"pinata": "^2.4.9",
+		"stripe": "^18.3.0"
 	}
 }
