+++ api/src/hooks.server.ts
@@ -1,0 +1,11 @@
+import { dev } from '$app/environment';
+import type { HandleFetch } from '@sveltejs/kit';
+import { APP_DOMAIN } from '../../shared/config';
+
+export const handleFetch: HandleFetch = async ({ request, fetch }) => {
+	const origin = request.headers.get('origin');
+	if (origin !== APP_DOMAIN && !dev) {
+		return new Response('Unauthorized', { status: 401 });
+	}
+	return fetch(request);
+};
