+++ api/src/lib/server/stripe.ts
@@ -1,0 +1,6 @@
+import Stripe from 'stripe';
+import { env } from '$env/dynamic/private';
+
+export const stripe = new Stripe(env.VITE_STRIPE_SECRET_KEY, {
+	apiVersion: '2025-08-27.basil'
+});
