--- app/src/lib/server/supabase.ts
@@ -1,4 +1,0 @@
-import { PUBLIC_SUPABASE_ANON_KEY, PUBLIC_SUPABASE_URL } from '$lib/global/config';
-import { createClient } from '@supabase/supabase-js';
-
-export const supabase = createClient(PUBLIC_SUPABASE_URL, PUBLIC_SUPABASE_ANON_KEY);
