+++ app/src/routes/+layout.server.ts
@@ -1,5 +1,6 @@
 import type { LayoutServerLoad } from './$types';
 import type { LikedTrackObject, UserProfile } from '../../../shared/types';
+import { API_BASE } from '$lib/global/config';
 
 export const load: LayoutServerLoad = async ({ locals: { safeGetSession }, cookies, fetch }) => {
 	const { session, user } = await safeGetSession();
@@ -8,10 +9,12 @@ export const load: LayoutServerLoad = async ({ locals: { safeGetSession }, cooki
 	let likedTracks: LikedTrackObject[] = [];
 	let followedIDs: string[] = [];
 
+	const userId = session?.user.id;
+
 	if (session) {
-		profileData = await fetch(`/api/users/${session.user.id}`).then((res) => res.json());
-		likedTracks = await fetch(`/api/users/${session.user.id}/likes`).then((res) => res.json());
-		followedIDs = await fetch(`/api/users/${session.user.id}/following`).then((res) => res.json());
+		profileData = await fetch(`${API_BASE}/users/${userId}`).then((res) => res.json());
+		likedTracks = await fetch(`${API_BASE}/users/${userId}/likes`).then((res) => res.json());
+		followedIDs = await fetch(`${API_BASE}/users/${userId}/following`).then((res) => res.json());
 	}
 
 	return {
