+++ app/src/routes/artists/+page.ts
@@ -1,7 +1,8 @@
+import { API_BASE } from '$lib/global/config';
 import type { ArtistRaw } from '../../../../shared/types';
 
 export const load = async ({ fetch }) => {
-	const artists: ArtistRaw[] = await fetch('/api/artists').then((res) => res.json());
+	const artists: ArtistRaw[] = await fetch(`${API_BASE}/artists`).then((res) => res.json());
 
 	return {
 		artists
