+++ app/src/routes/likes/+page.server.ts
@@ -1,4 +1,5 @@
 import type { Actions } from '@sveltejs/kit';
+import { API_BASE } from '$lib/global/config';
 
 export const actions: Actions = {
 	toggleLikedTrack: async ({ request, fetch, locals: { safeGetSession } }) => {
@@ -8,7 +9,7 @@ export const actions: Actions = {
 			const trackId = formData.get('trackId');
 			const addOrRemove = formData.get('addOrRemove');
 			if (trackId && addOrRemove) {
-				await fetch(`/api/users/${session.user.id}/likes`, {
+				await fetch(`${API_BASE}/users/${session.user.id}/likes`, {
 					method: addOrRemove === 'remove' ? 'DELETE' : 'POST',
 					headers: {
 						'Content-Type': 'application/json'
