+++ app/src/routes/releases/+page.ts
@@ -1,7 +1,8 @@
+import { API_BASE } from '$lib/global/config';
 import type { ReleaseHydrated } from '../../../../shared/types';
 
 export const load = async ({ fetch }) => {
-	const releases: ReleaseHydrated[] = await fetch('/api/releases').then((res) => res.json());
+	const releases: ReleaseHydrated[] = await fetch(`${API_BASE}/releases`).then((res) => res.json());
 
 	return {
 		releases
