+++ app/src/routes/releases/[slug]/+page.ts
@@ -1,7 +1,8 @@
+import { API_BASE } from '$lib/global/config';
 import type { ReleaseHydrated } from '../../../../../shared/types';
 
 export const load = async ({ fetch, params }) => {
-	const release: ReleaseHydrated = await fetch(`/api/releases/${params.slug}`).then((res) =>
+	const release: ReleaseHydrated = await fetch(`${API_BASE}/releases/${params.slug}`).then((res) =>
 		res.json()
 	);
