+++ app/vite.config.ts
@@ -1,11 +1,12 @@
 import { svelteTesting } from '@testing-library/svelte/vite';
 import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
+import { APP_LOCAL_PORT } from '../shared/config';
 
 export default defineConfig({
 	plugins: [sveltekit()],
 	server: {
-		port: 3000
+		port: APP_LOCAL_PORT
 	},
 	test: {
 		workspace: [
