+++ dashboard/vite.config.ts
@@ -1,6 +1,10 @@
-import { sveltekit } from '@sveltejs/kit/vite';
-import { defineConfig } from 'vite';
+import { sveltekit } from "@sveltejs/kit/vite";
+import { defineConfig } from "vite";
+import { DASHBOARD_LOCAL_PORT } from "../shared/config";
 
 export default defineConfig({
-	plugins: [sveltekit()]
+  plugins: [sveltekit()],
+  server: {
+    port: DASHBOARD_LOCAL_PORT,
+  },
 });
