+++ shared/config/index.ts
@@ -1,0 +1,26 @@
+export const APP_DOMAIN = "https://soli.network";
+export const APP_LOCAL_PORT = 3000;
+
+export const DASHBOARD_DOMAIN = "https://dashboard.soli.network";
+export const DASHBOARD_LOCAL_PORT = 3001;
+
+export const API_DOMAIN = "https://api.soli.network";
+export const API_LOCAL_PORT = 3002;
+
+export const REVENUE_SPLIT = {
+  artists: 0.9,
+  platform: 0.1,
+};
+
+export const TABLES = {
+  users: "users",
+  artists: "artists",
+  artistMembers: "artist_members",
+  tracks: "tracks",
+  releaseTracks: "release_tracks",
+  releasesHydrated: "hydrated_releases",
+  streams: "streams",
+  betaUsers: "beta-users",
+  likedTracks: "liked_tracks",
+  followedArtists: "followed_artists",
+};
